feat: add subtitle to collection item
